* Done with optimization process

+++ src/main/java/com/sbt/cdm/api/model/extension/PublishedHsDeposit.java
@@ -86,7 +86,7 @@ public class PublishedHsDeposit {
     @OraName(value = "H_PROC_COUNT")
     @InitOrder(value = "10")
     @DataType(value = TransformType.LONG)
-    public Long procCount;
+    public BigDecimal procCount;
 
     @OraName(value = "H_STATE")
     @InitOrder(value = "11")
